refactor: 레포지토리 test 패키지 구조 변경
