+++ src/test/java/qna/repository/QuestionRepositoryTest.java
@@ -1,11 +1,10 @@
-package qna.repository.question;
+package qna.repository;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import qna.domain.Question;
-import qna.repository.QuestionRepository;
 
 import java.util.List;
 import java.util.Optional;
