+++ src/test/java/qna/repository/UserRepositoryTest.java
@@ -1,11 +1,10 @@
-package qna.repository.user;
+package qna.repository;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import qna.domain.User;
-import qna.repository.UserRepository;
 
 import java.util.Optional;
